requirements.txt
python-telegram-bot==13.15
requests>=2.31
pytz>=2024.1
